Inject logging transformation function in Logger
This makes it easier to extend the Logger by extracting the logic to get
from `Any` to a `String` message into a pure and standalone function.

It also makes it easier to test the component. We need only one test for
the logging action now, and we can test the transformation function in
isolation.