Make Logger expect a Loggable input
This makes it easier to extend the Logger, as the thing required for a
something to be logged is for it to conform to Loggable.

It also makes it easier to test the behaviour. Logger only needs one
test, to ensure it uses its Loggable import properly.

Types conforming to Loggable can be tested in isolation, but it can be
argued testing a static string is not very valuable. The only exception
might be if you want to adda layer of safety to make sure changes to the
static value are intentional. An unintentional test would break the
tests.